pyproject.toml
[tool.poetry]
name = "final_project"
version = "0.1.0"
description = "environment for final project"
authors = ["eenzeenee <lililio0170@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
fastapi = "^0.89.1"
pytest = "^7.2.1"
torch = "^1.13.1"
transformer-in-transformer = "^0.1.2"
kss = "^4.4.0"
sentence-transformers = "^2.2.2"
tqdm = "^4.64.1"
pandas = "1.5.3"
nltk = "^3.8.1"
uvicorn = "0.15.0"
streamlit = "1.2.0"
protobuf = "3.19.6"
scikit-learn = "^1.2.0"
omegaconf = "^2.3.0"
pydub = "^0.25.1"
soundfile = "^0.11.0"
openai-whisper = "^20230117"
whisper = "^1.1.10"
numpy = "1.23.5"
librosa = "^0.9.2"
python-multipart = "^0.0.5"
ffmpeg-python = "^0.2.0"
python-mecab-ko = "^1.3.3"
gluonnlp = "0.10.0"
mxnet = "^1.9.1"
torchcrf = "^1.1.0"
pytorch-crf = "0.7.2"
konlpy = "^0.6.0"
tweepy = "3.10.0"
keybert = "0.7.0"
sentencepiece = "0.1.96"
keras = "2.9.0"
tensorflow = "2.9.3"
mecab = "0.996.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"